Refactor how userd keeps track of intercepts
This commit introduces a mutex protected `intercept` data structure
that combines information about an ongoing intercept, together with a
context and its cancellation. An instance of the data structure is
created when an intercept arrives (through the WatchIntercepts watcher)
and is cancelled when it departs.

A more fine-grained `podIntercept` (formerly known as `liveIntercept`)
still exists to keep track of the pod chosen for port-forwards and
remote mounts.

The `intercept` struct provides a cleaner way for the user daemon
to track the intercepts. The `TrafficManager` structure used to have
the following attributes:

  mountPoints             sync.Map
  mountMutexes            sync.Map
  currentIntercepts       []*manager.InterceptInfo
  currentInterceptors     map[string]int
  activeInterceptsWaiters sync.Map

This is now reduced to:

  currentIntercepts map[string]*intercept
  interceptWaiters map[string]*awaitIntercept

In addition to the above, the 90+ lines method
`workerPortForwardIntercepts` with the `//nolint:gocognit // bugger off`
comment was refactored into more readable functions.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>