Make sliceValue flag types implement pflag.SliceValue
The `cliutil.Value` implementation of `StringArrayVar`, `StringSliceVar`
etc. didn't implement the `pflag.SliceValue` interface. This commit
ensures that they do. The `complexValue` was renamed to `sliceValue` and
dedicated to `pflag.SliceValue` implementations (it is no longer used
for the "ipNet").

This commit also enforces consistent point receiver usage for all
`cliutil.Value` implementations.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>