Use sigs.k8s.io/yaml instead of gopkg.in/yaml.v3 for agent config
The `sigs.k8s.io/yaml` is sensitive to the `json:""` annotations in
a struct, and doesn't require them to be duplicated with `yaml:""`. It
also perform correct encoding/decoding of all Kubernetes types.

Unfortunately, it doesn't provide the knowledge about lines and
columns during parsing, which makes it unsuitable in some situations,
such as when parsing our config.yml.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>