Add ability to configure resources for the agent and init container
Adds the new Helm Chart values `resources` and `initResources` to the
`agentInjector.agentImage` structure in the Helm chart. The values
end up as JSON strings in the environment variables `AGENT_RESOURCES`
and `AGENT_INIT_RESOURCES` in the traffic-manager deployment, where
the traffic-manager picks them up and propagates them to all agent
configurations.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>