Add watcher that detects events related to agent injection problems
Adds a watcher for events generated by Kubernetes that filters out
problems that indicates that an injection attempt failed. The watcher
is added just before the agent configmap is stored, which triggers
the rollout and hence, the injection. If issues arise, they are
propagated to the user with a suggestion on how to proceed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>
Signed-off-by: Raphael Reyna <raphaelreyna@protonmail.com>