Run go-fusefpt via gRPC so that we don't require a fuse installation
If we link the telepresence binary with a fuse library, then that
library must be installed in order for telepresence to run. This is a
problem on platforms like macOS and Windows where the fuse installation
is either proprietary or GPL3, because we cannot bundle it with our
installation.

Running a linked binary via gRPC will not stop the telepresence binary
from functioning in case the gRPC server doesn't start.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>