Ensure that DNS is configured before returning from CreateIntercept
If an intercept is using a `--` flag to run a command, it's essential
that the DNS has been configured properly before that command is
started, or it will fail to resolve properly.

This commit adds a `waitForDNS` function that delays the return from the
call to `CreateIntercept` until the DNS has been configured.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>