Deprecate Helm chart dnsConfig in favor of client.routing
The `dnsConfig` entry in the Helm chart did not contain anything related
to DNS. The `alsoProxySubnets` and `neverProxySubnets` are both related
to how the client's virtual network interface configures its routing.

This commit retains the `dnsConfig` for backward compatibility, but
marks it as deprecated, and adds a new `client` struct with a new
`routing` struct that contains the also/never proxy entries.

The commit also renames the gRPC messages, but retains the type and
position to keep things backward compatible.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>