Don't dup stdio to logger
The change in 2.7.6 where stdout and stderr was duped to the logrus
only worked for normal IO. Things like panics, or stack trace output
from a QUIT signal, just caused everything to hang, only killable with
a `kill -9`.

This commit removes the extra step, and just dups the file descriptor
of the log file itself (back to how it was prior to 2.7.6).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>