Fix race condition when closing TUN-device
Closing the TUN-device would sometimes result in a panic and killing the
root daemon, because the stack would remove the attached dispatcher
just after a reading the packet from the TUN-device but before the call
to `InjectInbound`. This commit adds a mutex to protect the dispatcher
(the methods `Attach` and `IsAttached`) and makes the `IsAttached` the
only way to determine if the device is still running.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>