Fix environment variable interpolation in generated traffic-agent
Environment variables that are defined in a pod's container can be
interpolated in other values using a `$(<env name>)` syntax. This commit
ensures that such interpolation works with the env-vars that are copied,
and prefixed into the traffic agent by also prefixing the actual
interpolation. E.g.
```
volumeMounts:
  - name: $(TOKEN_VOLUME)
```
becomes:
```
volumeMounts:
  - name: $(_TEL_APP_A_TOKEN_VOLUME)
```
when copied into the agent container definition.

Closes #2756

Signed-off-by: Thomas Hallgren <thomas@datawire.io>