Add agent struct to the Helm chart
I wanted to add settings for the traffic-agent's Envoy admin and server
ports. Adding them to `agentInjector` or `agentInjector.agentImage` just
didn't feel right because it has nothing to do with them, so I added
a special `agent` struct instead, with the intention to capture things
related to the agent container runtime that isn't related to the
injector.

No Envoy settings are added in this commit.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>