Introduce datawire/envconfig and move environment parsing to one place
The `datawire/envconfig` replaces sethvargo/go-envconfig in this commit,
enabling us to unify all environment string parsing to one place. This
commit also changes the way defaults are handled so that all such values
are declared in the Helm chart, and only in the Helm chart. The
rationale for this is that some values are used in more places than the
environment, and would therefore need to have defaults in more than one
place unless they are in the Helm chart values.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>