Get rid of dnsServiceXXX settings in Helm chart
On some platforms, the telepresence client must dedicate an IP address
covered by one of the TUN device's CIDRs for the DNS resolver. The
chosen IP must not collide with anything relevant in the cluster.

Before this commit, the traffic-manager would find the address of the
cluster's DNS service, and the client would then use that. The rationale
being that no process in the client ought to access the cluster's DNS
anyway. This approach has proven problematic because the DNS service
name and namespace is different in different clusters. To solve that,
some additional Helm chart values have been provided to make it possible
to configure what service.namespace to use or event to hardcode the
actual IP. All of this is unnecessary.

There are two requirements on the IP-address; it must be covered by a
TUN-device CIDR and it must be irrelevant to otherwise access it. We
already have that IP. It's the Pod-IP of the traffic-manager.

So, long story short, let's use the traffic-manager's podIP and get rid
of all magic around the `dnsServerXXX` settings.

Closes #2760

Signed-off-by: Thomas Hallgren <thomas@datawire.io>