Change mutating webhook port to 443 and make it configurable.
Prior to this commit, the mutating webhook was listening to port `8443`.
The port was selected so that it was above the privileged port range,
and hence would cause no conflicts when running the container as an
unprivileged user. This caused problems with firewalls that enable port
`443` but blocks `8443` by default.

The remedy for this problem was to add:

    RUN setcap 'cap_net_bind_service+ep' /usr/local/bin/traffic

to the Dockerfile that builds the tel2 image. This enables the binary
to listen to privileged ports without being root.

This commit also fixes a bug in the Helm chart, causing the existing
`agentInjector.webhook.port` to be completely ignored. The default
setting for that value was incidentally already `443`, and now that
value is propagated correctly.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>