Have CanIntercept and AddIntercept use intercept instead of impl.
The `CanInterface` and `AddIntercept` are removed from the `Session`
interface. Instead, the `Session` interface is passed as a first
arguments to functions with the same name. This enables those functions
to call methods on the interface internally, instead of calling them
directly on the implementation.

It's a pity that interfaces cannot be used as call receivers in Go.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>