Never use an agent-image default.
This commit ensures that the traffic-manager can determine what image
to use for the traffic-agent, so that it never uses a default. The
image is either picked from the `AGENT_IMAGE` environment variable,
or if that is unset, retrieved from Ambassador Cloud.

If the Ambassador Cloud retrieval fails although access is configured,
it will be retried indefinitely, and injections and intercepts will be
disabled until the retrieval succeeds.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>