Handle missing cloud config.
If access to Ambassador Cloud was disabled, the traffic-manager would
return an "error while marshaling: proto: Marshal called with nil" error
from the gRPC method `GetCloudConfig`, which in turn terminated services
in the enhanced user daemon.

This commit ensures that the proper gRPC error code `Unavailable` is
returned with a descriptive message.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>