Fix panic when root daemon session creation times out.
Removes a redundant close of the `vifReady` channel. The close is
unnecessary and will sometimes cause a panic in `onFirstClusterInfo`,
because at the point where it happened, the cluster info watcher has
already been started, and `onFirstClusterInfo` will close the channel
when it receives the first cluster info. If that never happens, the
whole thing will time out and the session as a whole is cancelled.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>