Fix inconsistencies when `--output=json` is used.
Some output, like "An update of telepresence ...", "Launching
Telepresence Root Daemon", and "Connected to context ..." was printed
verbatim as unformatted text instead of being included in the JSON
output when the `--output=json` flag was used.

This commit ensures that information like that is discarded so that the
JSON output is always valid, and the "stdout" is kept consistent.

The commit also ensures that invalid `--output` flag values render a
proper error message, and that attempts to combine string output with
objects on formatted stdout renders a panic (since it's a coding error).

Closes #2854.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>