Watch traffic-manager-client configmap directly instead of mounting it.
Mounting a configmap and then using a file-watcher to get notified when
it changes is a very slow process (can take minutes). Watching the
configmap directly is however almost instant, and also a technique used
already by the agent-injector.

This commit replaces the mount of the configmap and the file-watcher
with a k8s watcher for the configmap.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>