Let root daemon have its own traffic-manager connection.
The root daemon uses a gRPC proxy exposed by the user daemon, which
means that every I/O operation performed on the tunnel that connects
the TUN-device to the traffic-manager is routed through the user daemon.
This has two disadvantages:

1. The extra I/O pump in another process degrades performance.
2. All gRPC streams are shared in the same underlying gRPC connection
   to the `ManagerClient`. gRPC has a limit of about 100 such streams
   per gRPC connection.

In this commit, the root daemon will create its own port forward
directly to the traffic manager, and then use that to connect a
`ManagerClient` of its own. This means that all outbound traffic is
routed directly to the traffic-manager without interfering with other
gRPC related tasks, or with inbound traffic.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>