Restore the ability to set loglevel in the root daemon.
The ability set the loglevel in the root daemon using the command
`telepresence loglevel <level>` got lost 2.8.0 rewrite, where the CLI no
longer communicates directly with the root daemon. The call was removed
from the CLI but never added to the user daemon.

This commit adds the missing `SetLogLevel` call from the user daemon to
the root daemon.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>