Make version output nicer, and add DisplayName extension point.
Replaces the key/value formatting of `describeIntercept` with a generic
`ioutil.KeyValueFormatter`, and also uses that formatte to format the
output of `telepresence version`.

The `cmd_version` command now also uses a global variable `DisplayName`,
which is intended to be an extension point so that `Client` can be
replaced with `Enhanced Client` or similar.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>