Make the integration_test/itest package reusable from other modules.
Changes the initialization so that a directories for the OSS root and
another module (the tested module) can be provided when the test harness
is created.

Also moves all k8s files that the tests depends on to `testdata/k8s`
beneath the `integration_test` to make it easy to track the manifests
used by the tests.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>