Node CIDRs are not pod CIDRs unless they cover the traffic-manager's IP
When using cluster's like Minikube with a Docker driver, the node CIDRs
are not used as pod CIDRs.

This commit checks if the IP of the traffic-manager is covered by a node
CIDR, and if not, the node strategy is abandoned. If the default "auto"
`PodCIDRStrategy` is used, that means that the discovery fall back to
"coverPodIPs".

Signed-off-by: Thomas Hallgren <thomas@datawire.io>