Fix panic when altering the config when no session exists.
Altering the `config.yml` file when after a `telepresence connect`
and a `telepresence quit` (i.e. when the daemons run but no session is
active) resulted in a panic in the user daemon. This commit fixes the
panic and also ensures that the config refresh looks the same in both
daemons.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>