Fix race condition when receiving multiple connect calls.
Sending rapid connect/leave requests to the daemons would sometimes
result in a hanging daemon because a caller would hang on the
session's done channel although no session was ever started.

This commit moves the start of the session into a method of its own
which makes it easier to read and control using a mutex.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>