Ensure that google credentials are picked up correctly by root daemon.
A user using `gcloud auth` to authenticate to a GKE cluster would not
be able to login unless the kubeconfig also contained a gce config
element. This was becaue the root daemon was looking in the wrong place
for the generated credentials file.

This commit ensures that setting the `GOOGLE_APPLICATION_CREDENTIALS`
environment variable has the desired effect, and also that the default
credentials are found when it is not set.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>