Let the Helm chart hooks include the helm revision as a query parameter.
A race condition may occur during upgrade where the http call issued
from a job started by a Helm pre-delete hook, is served by the upgraded
traffic-manager rather than the deleted one.

This commit adds the `Helm.Revision` as a `HELM_REVISION` environment
variable to the traffic-manager, and as a query parameter to the URL
calls that the Helm hooks perform. The service then compares those
revisions and discards the call when they don't match.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>