Add ability to get detailed formatted output from the intercept command.
Adds the boolean flag `--detailed-output` to be used together with
`--output=json|yaml`. When set, the output will be a formatted object
with all details of the intercept, including the environment.

This is intended to be used when running the CLI in a docker container
so that the environment and mount points can be captured and propagated
when starting another docker container, thereby removing the need to
pass information using a shared volume.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>