Add --local-mount-port option to the intercept command.
Adds a `telepresence intercept` flag `--local-mount-port <port number>`.
The user daemon will not perform mounts of remote directories when this
flag is given. Instead, it will create a bridge the given port on
localhost to the SFTP server running in the intercepted agent.

This will enable two things:
1. When the daemon runs in a Docker container, and in Docker Desktop
   specifically, the port can be published so that the host can do the
   actual mount without having access to the cluster network.
2. Any remote mounter capable of doing SFTP mounts, can be used.

Mounting by forwarding a port this way only works with SFTP because FTP
will try to open additional ports in the server when it does PASV/EPSV.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>