Make it possible to control the UID:GID used when creating cache files.
This commit alters the `dos` filesystem slightly so that the UID and
GID used when creating new files or directories can be controlled by
the environment variables `TELEPRESENCE_UID` and `TELEPRESENCE_GID`.

The addition is intended to prevent that cache files are created by root
when running the daemon as root in a docker container.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>