Add additional `--docker` flag to `telepresence connect`.
This commit moves the `telepresence connect` command into its own file.
It also adds the `connect.Request` struct as a wrapper around the rpc
based `ConnectRequest`. The wrapper can contain properties that are CLI
only (and hence not passed over to the user daemon). One such property
is the new `Docker` which controls if a daemon should be started in
a docker container or on the host.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>