Enable passing of `--context` to all commands.
Now when the context is used when identifying what daemon to connect
to, it makes a lot of sense to be able to send the context with all
commands that potentially connects to the daemon. This enables future
scenarios where multiple daemons running in different docker containers
can serve different contexts.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>