Fix status and version output when using a docker daemon.
When running the daemon remotely (i.e. in docker), there's only one
daemon. This commit ensures that the output from the status and version
commands show more accurate info. Instead of a "User daemon" and a
"Root daemon", it shows "Daemon in container" and then formatted data
beneath that heading.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>