Ensure that the kubeauth server restarts on configuration changes
If the `KUBECONFIG` environment or `--kubeconfig` flag changes, or if
any of the files referenced by them changes, then the kubeauth server
must be restarted.

This commit ensures this restart by storing the list of configFiles
in the portfile that the server produces, and by adding a watcher for
both the port file and for the config files. The server exists if a
change is detected.

The client checks if the list of config files are the same, and if not,
it removes the portfile and restarts the server.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>