Ensure that container based daemon quits on early failure.
A container based daemon must quit immediately if the session creation
fails, because it can only handle one single session. A failure to quit,
means a failure to stop the container, which in turn means that the
container keeps on running and that the next connect attempt then fails
because the daemon container handling the given context already exists.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>