Ensure correct connect response even when context is cancelled.
This commit fixes an intermittent problem where a context cancellation
would cause a response to never be posted on the `connectResponse`
channel, and hence cause a bad reply to the client in the form of
an EOF when the process instead died.

A command that often reproduced the problem was

    telepresence connect --docker

on a cluster where no traffic-manager was installed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>