Run docker daemon in detached mode.
If another process executes the command `telepresence connect --docker`
and then waits for the command completion, the wait will not return
although the telepresence process exits because the `docker run`
subprocess will keep `stdout` and `stderr` file descriptors open.

The remedy for this is to run the docker container in detached mode
using the `-d` flag.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>