Ensure that the docker container name is valid.
The name used for `docker run --name <name>` must comply with the
regular expression `[a-zA-Z0-9][a-zA-Z0-9_.-]`. This is however not true
for the name of a Kubernetes context. So when using a context name as
the argument to `--name`, it must be mangled so that any illegal
characters are replaced.

This commit adds the `SafeContainerName` function which replaces all
illegal characters with an underscore unless it's the very first
character, in which case it is replaced with an 'a'.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>