Enable control of test environment from a configuration file.
This addition is intended for users that want to run the integration
tests locally from an IDE without having to add environment variables
to the run configuration.

This commit adds a new `itest.yml` config file that can be placed
adjacent to the `config.yml` file. It currently contains only one
entry, the Env. Sample content:

    Env:
      DEV_TELEPRESENCE_VERSION: v2.12.1-alpha.1
      DTEST_KUBECONFIG: /home/thhal/Downloads/thallgren-1.26.yaml
      DTEST_REGISTRY: localhost:5000

Signed-off-by: Thomas Hallgren <thomas@datawire.io>