Let UserHomeDir panic instead of returning an error.
Basically nothing will work if the user's home directory cannot be
determined, yet we diligently (well almost) propagated that error in
a lot of places in the code without acting on it. In some places, we
simply left the error unchecked.

This commit changes this so that a panic is thrown, and removes all
error propagation.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>