Don't use named pipes in windows.
Microsoft's Go implementation of named pipes sometimes causes the
Telepresence daemons to hang indefinitely. This is the culprit:

https://github.com/microsoft/go-winio/issues/281

This commit removes the use of named pipes (once implemented due to
buggy Go implementation of unix sockets), in favor of using the same
socket implementation (almost) as we do on other platforms.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>