Ensure proper DNS function when no cluster CIDR is routed by the VIF.
Telepresence does not route cluster subnets that are already available
on the workstation, which is good, but the DNS malfunctions on Linux and
Windows unless there is no subnet covering the IP of the Telepresence
DNS server, which is bad.

This commit ensures that a subnet is routed by the VIF even when none of
the cluster's subnets are routed. In such situations, a tiny subnet is
created with a 30-bit mask. It is guaranteed not to collide with any of
the subnets used by existing network interfaces. The DNS server will
then use an IP in that subnet.

On darwin (macOS), no special DNS subnet is needed because the DNS is
managed by adding and removing entries under /etc/resolver that points
directly to a port on localhost. So when no subnets are routed, there's
actually no need for a VIF. This commit takes this into account, so that
no VIF is started unless there's a need for it.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>