Fix problem with the DNS recursion check.
Adds a short timeout and a repeat to the DNS recursion check so that
it doesn't need to timeout in the cluster. The rationale behind this is
that if the DNS indeed is recursive, then the cluster's DNS server must
have access to the network where Telepresence DNS server is present on
on the client, which means that the cluster must be on the client and
the DNS should therefor be very fast.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>