Replace search-path that isn't intended for the cluster DNS resolver.
This applies only to the overriding resolver that we use in Docker and
on Linux where systemd.resolved isn't configured.

Instead of adding a search path declared in a resolv.conf file to the
exclude-suffix list, this commit ensures that such paths are is stripped
off before the query is subjected to Telepresence search-path and sent
to the cluster.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>