Don't report errors from fallback resolver when there's a local hit
This commit addresses the problem where the Telepresence resolver finds
an entry for an A record, but not for an AAAA record, and the AAAA
query therefore is sent to the fallback resolver. In such situations,
the NXNAME error from the AAAA query must be replaced by an EMPTY
result, indicating that the A record is sufficient. The same is true
for the opposite where an AAAA record is found but the A is sent to
the fallback.

Fallback servers are only used by the overriding resolver.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>