Add configurable strategy to set the global DNS search path on Windows
The global DNS search path can be set, either by using Powershell or
directly, by setting a registry value. Using Powershell will sometimes
fail due to group policies and setting the registry value doesn't seem
to work on all Windows platforms.

This commit introduces a strategy setting that can be set to "auto",
"powershell", or "registry" where "auto" means that "powershell" is
tried first, and if it fails, fallback to "registry". The default
setting is "auto".

Signed-off-by: Thomas Hallgren <thomas@datawire.io>