Ensure that tunnels to a podIP is redirected to the correct pod.
Tunneling to a pod was previously made so that the dial made at the
tunnel endpoint was made from a pod in the intercepted namespace. This
isn't enough when using FTP in static mode, because the FTP server will
check that `data` connections originate from the same IP as the original
`control` connection. If they differ, the `data` connection is rejected.
This meant that data transfers in FTP mounted file systems would fail
when a client had several simultaneous intercepts.

This commit ensures that if the destination IP is the same as the podIP
of an intercept, then the tunnel is redirected to that pod so that the
dial is made there.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>