Wait for unmount to complete.
An unmount of a file system may happen in situations where large amounts
of data has just been written to the local filesystem, which means that
a sync is in progress to transfer this data to the remote filesystem.
It's important that such syncs are allowed to complete before the
unmount operation finishes.

This commit adds the needed wait-groups and also adds two configurable
timeouts to the fuseftp mounter.

    timeouts:
      # FTP client shutdown grace period
      ftpReadWrite: <duration>

      # FTP client shutdown grace period
      ftpShutdown: <duration>

Signed-off-by: Thomas Hallgren <thomas@datawire.io>