Make UpdateCheckIfDue failure non-fatal.
When starting several telepresence intercepts in parallel, that
sometimes causes the update check file to be corrupt because the writes
aren't concurrency protected when several processes are writing at
the same time. As a result, the file containing the next timestamp may
be corrupted.

After this commit, such a file will be treat treated as empty instead of
generating an error. Maximum damage, in the unlikely event that a user
starts multiple intercepts at the same time, is that an extra update
check is made.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>