Add integration tests that reads and writes lots of data in parallel.
These tests are intended to test the stability by using several
intercepts that have a remote persistent store. The intercepts are
started, data is written, then the intercepts are stopped. They are then
started again, and the previously written data is validated at the same
time as new data is written. Then the intercepts are stopped and
restarted again, and all data is validated.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>