Don't let failure to start the fuseftp-server stop the daemon.
The start of the fuseftp server is deferred to the moment when a
mount is requested that requires it, which is good. However, when fuse
isn't available, the attempt to start the server returns an error, and
that error was allowed to stop the user daemon. This commit ensures that
this error instead is logged, and that the user daemon continues to
operate.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>