Better discovery of local container-based clusters (kind or minikube).
When using `telepresence connect --docker` in combination with a
kubeconfig that points to a cluster that runs in a container on the
local host, that kubeconfig must be modified to instead point to a
docker internal address:port, and the network that makes that address
available must be added to the Telepresence daemon container.

This commit improves the discovery of such configuration so that less
assumpions are made regarding container names and port numbers. Instead,
the discovery looks at some well-known labels and uses the docker
configurations port bindings to figure out how a port on the host maps
to a port in the container.

Closes #3155

Signed-off-by: Thomas Hallgren <thomas@datawire.io>