IPv6 enable the telepresence docker network.
DNS AAAA queries will never reach a Telepresence daemon that runs in a
Docker container unless that container is attached to an IPv6 enabled
network. This commit therefore ensures that the "telepresence" network
is IPv6 enabled by first trying to create it that way, and if that fails
due to the lack of an IPv6 pool in `/etc/docker/daemon.json`, by adding
a random IPv6 subnet and gateway to force the enablement without the
existence of such a pool.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>