Ensure that the docker client uses the current docker context.
The Docker client API has no knowledge of a docker context, so in this
commit, a `docker context inspect` is used to find out the docker host
of the current context. That host is then passed to the client API
instead of the default host so that the client API speaks to the host
of the current context, just like the docker binary does.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>