Use Docker's client API where possible instead of shelling out.
Replaces several shell-outs to docker with calls to the API. The pull,
build, and run were left intact. The pull needs docker registry
credentials which is somewhat cumbersome to pick up, and the build
and run commands parses a large number of flags that we currently don't
want to deal with.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>