Make Envoy's http idle timout configurable.
This timeout defaults to one hour, which means that connections will
often stay open that long. This is not ideal for intercepted connections
because they occupy resources all the way from the agent to the client.

This commit makes the timeout configurable and sets the default to 70
seconds. The default was chosen to play well with TCP keepalive polls
which often have a frequency of 60 seconds or less.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>