Ensure that it is the creators context that controls the bidi-pipe.
A connection that is created from a client to an agent, or vise versa,
starts when the initiator calls `Tunnel` which in turn calls
`EstablishBidiPipe`. That method then asks the peer to also call
`Tunnel`, which in turn calls `OnConnect` to attach the other end of
the bidi-pipe. Prior to this commit, it was the context of the caller
to `OnConnect` that controlled the bidi-pipe cancellation. This was bad,
because the bidi-pipe would then sometimes linger after close had been
called at the initiator end, leading to resource leaks. This commit
changes this so that it is instead the context of caller to
`EstablishBidiPipe` (the initiator) that controls the bidi-pipe.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>