Make traffic manager's state private and extendable through interface.
The former `State` struct is no longer exported after this commit.
Instead, the `State` can be accessed using an interface. The `state`
structure knows about this interface through the variable `self` which
makes it possible for itself to call on extended methods on itself.

Some very old tests were rewritten so that `oldtestcompat_test.go` could
be removed.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>