Pass the scout.Reporter in the context.
A lot of structures kept a ``*scout.Reporter` entry that was passed in
as a constructor parameter. This commit removes all of that in favor
of using a context value. This makes it easier to configure or mock the
reporter, and also to completely disable it, should we ever want to do
that. It also makes the reporter available everywhere there's a context.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>