Simplify the connector w.r.t. namespace.
The namespace is connected to one single namespace at a time, and this
also controls the DNS of the root daemon. In this commit, this namespace
is made immutable in the session, so the user must disconnect and then
connect again in order to switch the namespace.

Also, using the `--local-only` intercept flag is now a no-op if the
given namespace matches the namespace of the connection, or an error
if that's not the case.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>