Add global option --use <regexp> to identify a running daemon.
This commit introduces a new `daemon.Identifier` to identify a running
containerized daemon using the name of its kubecontext and the namespace
that it is connected to.

The commit also adds a global `--use <regex>` option that can be used
to identify what daemon a CLI command should use when there are several
daemons running.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>