Simplify consumption metrics collection.
The approach of using a channel to increment a counter would sometimes
result in the channel being filled up, because a depart could arrive
although data remained to be sent. The result would then be a very
large number of unhelpful errors saying:

    timeout trying to increment probe channel

This commit simplifies the metrics collection so that instead of using
a channel, it just uses an `atomic.AddUint64` and `atomic.LoadUint64`,
thereby removing any chance of congestion.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>