Introduce connection name.
Adds a `--name <name>` option to the `telepresence connect` command, so
that a connection can be given a name that is different from the
<kubeContext>-<namespace> name that is generated by default.

When a connection is named, a containerized daemon will be generated
by prepending "tp-" to that name.

The introduction of named connections enable use of multiple daemonized
connections that uses the same namespace. This is beneficial because
then pods with conflicting localhost ports can be intercepted
simultaneously into different containers (each with a unique localhost).

Signed-off-by: Thomas Hallgren <thomas@datawire.io>