Add explicit environment propagation between client and connector
This commit removes the code that passed environment variables in as
kubernetes flags, to instead pass them explicitly where needed.

Also adds propagation of the kubernetes flags to the client when it
reconnects, so that those flags can be used for other purposes.

The commit removes the propagation of kubernetes flags to the root
daemon. It has no use for them.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>