Don't clobber host's kubeconfig when connecting to containerized daemon.
Connecting to a containerized daemon means that we also modify the
`kubeconfig` flag to point to a copy of the config file in the cache.
This modification is now kept in a special map that serves as an
override when connecting. This means that the original host kubeconfig
flag (if any) isn't affected.

Signed-off-by: Thomas Hallgren <thomas@datawire.io>